fix(gates): split the unresolved-new test out of lower/tests.rs for the file cap
